requirements.txt
langchain
langchain-community
ollama
pyyaml
requests